Refactor Citations component into a folder for decomposition.
