+++ client/src/components/Citation/Citations.vue
